+++ scripts/scripts-dev.js
@@ -88,29 +88,6 @@ export function moveInstrumentation(from, to) {
 }
 
 /**
- * SalesForce MCP
- */
-
-function loadEvergageScript() {
-  const script = document.createElement('script');
-  if (window.location.host === 'lifesciences.danaher.com') {
-    script.src = 'https://cdn.evgnet.com/beacon/v55685555553mx3rf3h3n3n3i091550196/danaher_ls_prod/scripts/evergage.min.js';
-  } else {
-    script.src = 'https://cdn.evgnet.com/beacon/v55685555553mx3rf3h3n3n3i091550196/danaher_ls_staging/scripts/evergage.min.js';
-  }
-  script.onload = function onEvergageLoad() {
-  };
-  script.onerror = function onEvergageError() {
-  };
-  document.head.appendChild(script);
-}
-
-/* eslint-disable no-console */
-if (OnetrustActiveGroups.includes('C0004')) {
-  loadEvergageScript();
-}
-
-/**
  * Get the Image URL from Scene7 and Optimize the picture
  * @param {string} imageUrl
  * @param {string} imageAlt
